+++ packages/components-react/src/examples/SelectExampleAsyncFilter.tsx
@@ -4,6 +4,7 @@ import {
   PInputSearch,
   PSelect,
   PSelectOption,
+  type PSelectProps,
   type SelectChangeEventDetail,
   type SelectToggleEventDetail,
 } from '@porsche-design-system/components-react';
@@ -18,7 +19,7 @@ const useDebounce = <T,>(callback: (value: T) => void, delay = 400) => {
 };
 
 export const SelectExampleAsyncFilter = () => {
-  const [value, setValue] = useState<string | undefined>(undefined);
+  const [value, setValue] = useState<PSelectProps['value']>(undefined);
   const [options, setOptions] = useState<{ value: string; label: string }[]>([]);
 
   const [searchValue, setSearchValue] = useState('');