+++ packages/components-react/src/routes.tsx
@@ -81,6 +81,16 @@ export const routes: RouteType[] = [
     element: <fromExamples.AGGridExampleStorefrontPage />,
   },
   {
+    name: 'AG Grid Compact Example',
+    path: '/aggrid-compact-example',
+    element: <fromExamples.AGGridCompactExamplePage />,
+  },
+  {
+    name: 'AG Grid Compact Example Storefront',
+    path: '/aggrid-compact-example-storefront',
+    element: <fromExamples.AGGridCompactExampleStorefrontPage />,
+  },
+  {
     name: 'Banner Example',
     path: '/banner-example',
     element: <fromExamples.BannerExamplePage />,