+++ packages/components-vue/src/examples/InputTelExampleControlled.vue
@@ -1,8 +1,13 @@
 <script setup lang="ts">
-import { type InputTelInputEventDetail, PInputTel, PText } from '@porsche-design-system/components-vue';
+import {
+  type InputTelInputEventDetail,
+  PInputTel,
+  type PInputTelProps,
+  PText,
+} from '@porsche-design-system/components-vue';
 import { computed, ref } from 'vue';
 
-const value = ref<string>('');
+const value = ref<PInputTelProps['value']>('');
 
 const onInput = (e: CustomEvent<InputTelInputEventDetail>) => {
   value.value = (e.target as HTMLInputElement).value;